+++ GameMaker/src/edu/elon/honors/price/maker/MapEditorLayerActors.java
@@ -27,14 +27,14 @@ public class MapEditorLayerActors extends MapEditorLayer {
 		Tileset tileset = game.getMapTileset(map);
 		int tileHeight = tileset.tileHeight;
 		Bitmap bmp = parent.actorImage;
-		return (int)Math.round((touchY - parent.offY - bmp.getHeight() / 2) / tileHeight);
+		return Math.round((touchY - parent.offY - bmp.getHeight() / 2) / tileHeight);
 	}
 
 	private int getBitmapCol(float touchX) {
 		Tileset tileset = game.getMapTileset(map);
 		int tileWidth = tileset.tileWidth;
 		Bitmap bmp = parent.actorImage;
-		return (int)Math.round((touchX - parent.offX - bmp.getWidth() / 2) / tileWidth);
+		return Math.round((touchX - parent.offX - bmp.getWidth() / 2) / tileWidth);
 	}
 
 	@Override
